Added BlazorNavigationManagerExtensions
BlazorNavigationManagerExtensions is a simple set of extensions for the Navigation Manager to simplify its usage for things we are likely to do in most of our Blazor applications. It currently supports an overload to the NavigateTo method where you can specify query string parameters and a Reload method that force loads the current page. The goal would be to add methods that people need so that all the needs for the Navigation Manager are in one place.